refactor: move `hash_image` to `image_utils.py`
