+++ src/vidore_benchmark/utils/__init__.py
@@ -1,4 +1,5 @@
 from .data_utils import ListDataset
+from .image_utils import hash_image
 from .iter_utils import batched, islice
 from .logging_utils import setup_logging
 from .torch_utils import get_torch_device, tear_down_torch
