+++ src/vidore_benchmark/utils/image_utils.py
@@ -1,0 +1,16 @@
+import hashlib
+
+from PIL import Image
+
+
+def hash_image(image: Image.Image) -> str:
+    """
+    Hash a PILLOW image using MD5.
+
+    Args:
+        image (Image.Image): PIL Image object.
+
+    Returns:
+        str: MD5 hash of the image.
+    """
+    return hashlib.md5(image.tobytes()).hexdigest()
